fix:dev: factor out command implementation for `stack`, `cluster`, `hadoop`
Due to the "magic" used to call commands from a shell, there are
potential side-effects.  Since the `cm.CloudmeshConsole` dynamically
adds the methods of each plugin to itself, one command can shadow the
definition of a method or attribute.

This change pulls out the implementation methods for the `stack`,
`cluster`, and `hadoop` commands into a separate classes that is
instantiated within the `do_ZZZ` method of the plugin.